docker_submission/conic_template/requirements.txt
scikit-learn>=0.21
scikit-image
scipy>=1.2.1
opencv-python==4.5.1.48
pandas
tqdm
itk
# mahotas
segmentation_models_pytorch
nvidia-pyindex
nvidia-tensorflow
absl-py
dask
git+https://github.com/Kainmueller-Lab/evaluate-instance-segmentation.git@master#egg=evaluate-instance-segmentation
# git+https://github.com/Kainmueller-Lab/funlib.learn.tensorflow
git+https://github.com/Kainmueller-Lab/gunpowder.git@3ec561099ca70a352afd5246af31bfe6a6765eab#egg=gunpowder
git+https://github.com/maisli/neurolight.git@master#egg=neurolight
h5py
joblib
natsort
toml
zarr
colorcet
# Cython
dask
h5py
joblib
# mahotas
# malis @ git+https://github.com/TuragaLab/malis.git@master#egg=malis
matplotlib
neurolight @ git+https://github.com/maisli/neurolight.git@master#egg=neurolight
networkx
numcodecs
pycuda
pynrrd
python-louvain
tifffile
toolz
